+++ README.md
@@ -11,6 +11,12 @@
 ### Command
 <code>/maze</code>
 
+### Algorithms
+- A*
+- Best first search
+- Breadth first search
+- Depth first search
+
 ### More demos and walkthroughs
 <details>
 <summary><b>Initializing maze example</b></summary>
