+++ src/com/gubertmc/plugin/main/algorithms/breadthfirstsearch/bfs2d/BreadthFirstSearchSimulation2D.java
@@ -1,14 +1,15 @@
-package com.gubertmc.plugin.main.algorithms.dfs.dfs2d;
+package com.gubertmc.plugin.main.algorithms.breadthfirstsearch.bfs2d;
 
 import com.gubertmc.plugin.main.algorithms.Simulation;
 
-import java.util.Stack;
+import java.util.LinkedList;
+import java.util.Queue;
 
-public class DepthFirstSearchSimulation2D extends Simulation {
+public class BreadthFirstSearchSimulation2D extends Simulation {
 
     private boolean[][][] visited;
 
-    public DepthFirstSearchSimulation2D(int[][][] maze, int[] startCoordinate, int[] endCoordinate) {
+    public BreadthFirstSearchSimulation2D(int[][][] maze, int[] startCoordinate, int[] endCoordinate) {
         super(maze, startCoordinate, endCoordinate, false);
     }
 
@@ -26,16 +27,16 @@ public class DepthFirstSearchSimulation2D extends Simulation {
 
     @Override
     public boolean start() {
-        Stack<String> stack = new Stack<>();
-        stack.push(getStartCoordinate()[1] + "," + getStartCoordinate()[0]);
+        Queue<String> queue = new LinkedList<>();
+        queue.add(getStartCoordinate()[1] + "," + getStartCoordinate()[0]);
 
         boolean pathFound = false;
-        while (!stack.empty()) {
+        while (!queue.isEmpty()) {
             if (getTileGrid()[getEndCoordinate()[1]][getEndCoordinate()[0]][0] == 2) {
                 pathFound = true;
                 break;
             }
-            String x = stack.pop();
+            String x = queue.remove();
             int row = Integer.parseInt(x.split(",")[0]);
             int col = Integer.parseInt(x.split(",")[1]);
 
@@ -48,10 +49,10 @@ public class DepthFirstSearchSimulation2D extends Simulation {
             tileGrid[row][col][0] = 2;
             setTileGrid(tileGrid);
 
-            stack.push(row + "," + (col - 1)); //go left
-            stack.push(row + "," + (col + 1)); //go right
-            stack.push((row - 1) + "," + col); //go up
-            stack.push((row + 1) + "," + col); //go down
+            queue.add(row + "," + (col - 1)); //go left
+            queue.add(row + "," + (col + 1)); //go right
+            queue.add((row - 1) + "," + col); //go up
+            queue.add((row + 1) + "," + col); //go down
         }
         return pathFound;
     }
