+++ src/com/gubertmc/plugin/main/algorithms/breadthfirstsearch/bfs3d/BreadthFirstSearchAnimation3D.java
@@ -1,4 +1,4 @@
-package com.gubertmc.plugin.main.algorithms.dfs.dfs3d;
+package com.gubertmc.plugin.main.algorithms.breadthfirstsearch.bfs3d;
 
 import com.gubertmc.MazeGeneratorPlugin;
 import com.gubertmc.plugin.main.algorithms.Animation;
@@ -7,14 +7,15 @@ import org.bukkit.Material;
 import org.bukkit.scheduler.BukkitRunnable;
 
 import java.util.ArrayList;
-import java.util.Stack;
+import java.util.LinkedList;
+import java.util.Queue;
 
-public class DepthFirstSearchAnimation3D extends Animation {
+public class BreadthFirstSearchAnimation3D extends Animation {
 
     private boolean[][][] visited;
     private int[][][] textGrid;
 
-    public DepthFirstSearchAnimation3D(
+    public BreadthFirstSearchAnimation3D(
             MazeGeneratorPlugin plugin,
             Location[][][] tiles,
             int[] startCoordinate,
@@ -69,18 +70,18 @@ public class DepthFirstSearchAnimation3D extends Animation {
         int length = textGrid[0].length;
         int zLength = textGrid[0][0].length;
 
-        Stack<String> stack = new Stack<>();
-        stack.push(getStartCoordinate()[1] + "," + getStartCoordinate()[0] + "," + getStartCoordinate()[2]);
+        Queue<String> queue = new LinkedList<>();
+        queue.add(getStartCoordinate()[1] + "," + getStartCoordinate()[0] + "," + getStartCoordinate()[2]);
 
-        while (!stack.empty()) {
+        while (!queue.isEmpty()) {
             if (textGrid[getEndCoordinate()[1]][getEndCoordinate()[0]][getEndCoordinate()[2]] != 5) {
                 break;
             }
 
-            String x = stack.pop();
-            int row = Integer.parseInt(x.split(",")[0]);
-            int col = Integer.parseInt(x.split(",")[1]);
-            int z = Integer.parseInt(x.split(",")[2]);
+            String xs = queue.remove();
+            int row = Integer.parseInt(xs.split(",")[0]);
+            int col = Integer.parseInt(xs.split(",")[1]);
+            int z = Integer.parseInt(xs.split(",")[2]);
 
             if (row < 0 || col < 0 || row >= height || col >= length
                     || z < 0 || z >= zLength || visited[row][col][z]) {
@@ -93,12 +94,12 @@ public class DepthFirstSearchAnimation3D extends Animation {
             exploredPlaces.add(getTileGrid()[row][col][z]);
             setExploredPlaces(exploredPlaces);
 
-            stack.push(row + "," + (col - 1) + "," + z);
-            stack.push(row + "," + (col + 1) + "," + z);
-            stack.push((row - 1) + "," + col + "," + z);
-            stack.push((row + 1) + "," + col + "," + z);
-            stack.push(row + "," + col + "," + (z - 1));
-            stack.push(row + "," + col + "," + (z + 1));
+            queue.add(row + "," + (col - 1) + "," + z);
+            queue.add(row + "," + (col + 1) + "," + z);
+            queue.add((row - 1) + "," + col + "," + z);
+            queue.add((row + 1) + "," + col + "," + z);
+            queue.add(row + "," + col + "," + (z - 1));
+            queue.add(row + "," + col + "," + (z + 1));
         }
         return true;
     }
