+++ src/com/gubertmc/plugin/main/algorithms/depthfirstsearch/dfs2d/DepthFirstSearchSimulation2D.java
@@ -1,15 +1,14 @@
-package com.gubertmc.plugin.main.algorithms.bfs.bfs2d;
+package com.gubertmc.plugin.main.algorithms.depthfirstsearch.dfs2d;
 
 import com.gubertmc.plugin.main.algorithms.Simulation;
 
-import java.util.LinkedList;
-import java.util.Queue;
+import java.util.Stack;
 
-public class BreadthFirstSearchSimulation2D extends Simulation {
+public class DepthFirstSearchSimulation2D extends Simulation {
 
     private boolean[][][] visited;
 
-    public BreadthFirstSearchSimulation2D(int[][][] maze, int[] startCoordinate, int[] endCoordinate) {
+    public DepthFirstSearchSimulation2D(int[][][] maze, int[] startCoordinate, int[] endCoordinate) {
         super(maze, startCoordinate, endCoordinate, false);
     }
 
@@ -27,16 +26,16 @@ public class BreadthFirstSearchSimulation2D extends Simulation {
 
     @Override
     public boolean start() {
-        Queue<String> queue = new LinkedList<>();
-        queue.add(getStartCoordinate()[1] + "," + getStartCoordinate()[0]);
+        Stack<String> stack = new Stack<>();
+        stack.push(getStartCoordinate()[1] + "," + getStartCoordinate()[0]);
 
         boolean pathFound = false;
-        while (!queue.isEmpty()) {
+        while (!stack.empty()) {
             if (getTileGrid()[getEndCoordinate()[1]][getEndCoordinate()[0]][0] == 2) {
                 pathFound = true;
                 break;
             }
-            String x = queue.remove();
+            String x = stack.pop();
             int row = Integer.parseInt(x.split(",")[0]);
             int col = Integer.parseInt(x.split(",")[1]);
 
@@ -49,10 +48,10 @@ public class BreadthFirstSearchSimulation2D extends Simulation {
             tileGrid[row][col][0] = 2;
             setTileGrid(tileGrid);
 
-            queue.add(row + "," + (col - 1)); //go left
-            queue.add(row + "," + (col + 1)); //go right
-            queue.add((row - 1) + "," + col); //go up
-            queue.add((row + 1) + "," + col); //go down
+            stack.push(row + "," + (col - 1)); //go left
+            stack.push(row + "," + (col + 1)); //go right
+            stack.push((row - 1) + "," + col); //go up
+            stack.push((row + 1) + "," + col); //go down
         }
         return pathFound;
     }
